+++ example/lib/main.dart
@@ -1,39 +1,25 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_dialpad/flutter_dialpad.dart';
+import 'package:flutter_dialpad_example/screens/dialpad_selection.dart';
 
-
-void main() => runApp(MyApp());
-
-class MyApp extends StatefulWidget {
-  @override
-  _MyAppState createState() => _MyAppState();
+void main() {
+  WidgetsFlutterBinding.ensureInitialized();
+  runApp(MyApp());
 }
 
-class _MyAppState extends State<MyApp> {
-
+class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: Scaffold(
-        backgroundColor: Colors.black,
-        body: SafeArea(
-            child:
-            DialPad(
-                enableDtmf: true,
-                //outputMask: "(000) 000-0000",
-                hideSubtitle: false,
-                backspaceButtonIconColor: Colors.red,
-                buttonTextColor: Colors.white,
-                dialOutputTextColor: Colors.white,
-                keyPressed: (value){
-                  print('$value was pressed');
-                },
-                makeCall: (number){
-                  print(number);
-                }
-            )
+      theme: ThemeData(
+        iconTheme: IconThemeData(size: 36),
+        textButtonTheme: TextButtonThemeData(
+          style: ButtonStyle(
+            textStyle: MaterialStateProperty.all(TextStyle(fontSize: 20)),
+          ),
         ),
       ),
+      debugShowCheckedModeBanner: false,
+      home: DialPadSelection(),
     );
   }
 }