+++ lib/flutter_dialpad.dart
@@ -1,323 +1,326 @@
 library flutter_dialpad;
 
-import 'dart:async';
+import 'dart:math';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_masked_text2/flutter_masked_text2.dart';
-import 'package:flutter_dtmf/dtmf.dart';
+// disabled temporarily
+// import 'package:flutter_dtmf/dtmf.dart';
+
+import 'src/flutter_dialpad.dart';
+
+export 'src/flutter_dialpad.dart';
+
+typedef DialPadButtonBuilder = Widget Function(BuildContext context, int index, KeyValue key, KeyValue? altKey, String? hint);
 
 class DialPad extends StatefulWidget {
+  /// Callback when the dial button is pressed.
   final ValueSetter<String>? makeCall;
-  final ValueSetter<String>? keyPressed;
-  final bool? hideDialButton;
-  final bool? hideSubtitle;
-  // buttonColor is the color of the button on the dial pad. defaults to Colors.gray
-  final Color? buttonColor;
-  final Color? buttonTextColor;
-  final Color? dialButtonColor;
-  final Color? dialButtonIconColor;
-  final IconData? dialButtonIcon;
-  final Color? backspaceButtonIconColor;
-  final Color? dialOutputTextColor;
-  // outputMask is the mask applied to the output text. Defaults to (000) 000-0000
-  final String? outputMask;
-  final bool? enableDtmf;
 
-  DialPad(
-      {this.makeCall,
-      this.keyPressed,
-      this.hideDialButton,
-      this.hideSubtitle = false,
-      this.outputMask,
-      this.buttonColor,
-      this.buttonTextColor,
-      this.dialButtonColor,
-      this.dialButtonIconColor,
-      this.dialButtonIcon,
-      this.dialOutputTextColor,
-      this.backspaceButtonIconColor,
-      this.enableDtmf});
+  /// Callback when a key is pressed.
+  final ValueSetter<String>? keyPressed;
+
+  /// Whether to hide the dial button. Defaults to false.
+  final bool hideDialButton;
+
+  /// Whether to hide the backspace button. Defaults to false.
+  final bool hideBackspaceButton;
+
+  /// Whether to hide the subtitle on the dial pad buttons. Defaults to [false].
+  final bool hideSubtitle;
+
+  /// buttonColor is the color of the button on the dial pad. defaults to [Colors.grey]
+  final Color buttonColor;
+
+  /// Color of the button text, defaults to [Colors.black]
+  final Color buttonTextColor;
+
+  /// Color of the dial button, defaults to [Colors.green]
+  final Color dialButtonColor;
+
+  /// Color of the dial button icon, defaults to [Colors.white]
+  final Color dialButtonIconColor;
+
+  /// Icon for the dial button, defaults to [Icons.phone]
+  final IconData dialButtonIcon;
+
+  /// Color of the backspace button icon, defaults to [Colors.grey]
+  final Color backspaceButtonIconColor;
+
+  /// Color of the output text, defaults to [Colors.black]
+  final Color dialOutputTextColor;
+
+  /// Font size for the output text, defaults to 75
+  /// Text scales with the screen size using the shortest between (height & width) * 0.001 multiplied by the [dialOutputTextSize] value
+  final double dialOutputTextSize;
+
+  /// Font size for the output text, defaults to 75
+  final double buttonTextSize;
+
+  /// Font size for the output text, defaults to 25
+  /// Text scales with the screen size using the shortest between (height & width) * 0.001 multiplied by the [subtitleTextSize] value
+  final double subtitleTextSize;
+
+  /// outputMask is the mask applied to the output text. Defaults to (000) 000-0000
+  final String outputMask;
+  final String hint;
+
+  /// Whether to enable DTMF tones. Defaults to [false].
+  /// Disabled temporarily until [flutter_dtmf] has been updated.
+  final bool enableDtmf;
+
+  /// Builder for the keypad buttons. Defaults to [DialPadButtonBuilder].
+  /// Note: this has not yet been fully integrated for customer use - this will be available in a future release.
+  final DialPadButtonBuilder? keypadButtonBuilder;
+
+  /// Generator for the keypad buttons. Defaults to [PhoneKeypadGenerator].
+  final KeypadIndexedGenerator? generator;
+
+  /// Button display style (clipping). Defaults to [ButtonType.rectangle].
+  final ButtonType buttonType;
+
+  /// Padding around the button. Defaults to [EdgeInsets.all(0)].
+  final EdgeInsets buttonPadding;
+
+  /// Whether to call [makeCall] when the enter key is pressed. Defaults to false.
+  final bool callOnEnter;
+
+  /// Whether to copy the text to the clipboard when the text field is tapped. Defaults to true.
+  final bool copyToClipboard;
+
+  /// Whether to paste the text from the clipboard when the text field is tapped. Defaults to true.
+  final bool pasteFromClipboard;
+
+  /// Padding around the text field. Defaults to [EdgeInsets.only(left: 16, right: 16, top: 20, bottom: 16)].
+  final EdgeInsets textFieldPadding;
+
+  DialPad({
+    this.makeCall,
+    this.keyPressed,
+    this.hideDialButton = false,
+    this.hideBackspaceButton = false,
+    this.hideSubtitle = false,
+    this.outputMask = '(000) 000-0000',
+    this.hint = '(000) 000-0000',
+    this.buttonColor = Colors.grey,
+    this.buttonTextColor = Colors.black,
+    this.dialButtonColor = Colors.green,
+    this.dialButtonIconColor = Colors.white,
+    this.dialButtonIcon = Icons.phone,
+    this.dialOutputTextColor = Colors.black,
+    this.dialOutputTextSize = 75,
+    this.buttonTextSize = 75,
+    this.subtitleTextSize = 25,
+    this.backspaceButtonIconColor = Colors.grey,
+    this.enableDtmf = false,
+    @Deprecated('This has not yet been fully integrated for customer use and thus has no effect on the output - will be available in a future release.')
+    this.keypadButtonBuilder,
+    this.generator = const PhoneKeypadGenerator(),
+    this.buttonType = ButtonType.rectangle,
+    this.buttonPadding = const EdgeInsets.all(0),
+    this.callOnEnter = false,
+    this.copyToClipboard = true,
+    this.pasteFromClipboard = true,
+    this.textFieldPadding = const EdgeInsets.only(left: 16, right: 16, top: 20, bottom: 16),
+  });
+
+  /// Returns a [DialPad] with an iOS-style design (i.e. Apple).
+  factory DialPad.ios({
+    ValueSetter<String>? makeCall,
+    ValueSetter<String>? keyPressed,
+  }) {
+    return DialPad(
+      makeCall: makeCall,
+      keyPressed: keyPressed,
+      // Cupertino icons should be used here
+      dialButtonIcon: Icons.phone,
+      backspaceButtonIconColor: Colors.grey,
+      generator: IosKeypadGenerator(),
+      dialOutputTextColor: Colors.black87,
+      buttonTextColor: Colors.black87,
+      buttonColor: Colors.grey[300]!,
+      buttonType: ButtonType.circle,
+      buttonPadding: EdgeInsets.all(16),
+      dialOutputTextSize: 75,
+    );
+  }
+
+  /// Creates a dial pad with a metro-style (i.e. Windows).
+  factory DialPad.metro({
+    ValueSetter<String>? makeCall,
+    ValueSetter<String>? keyPressed,
+  }) {
+    return DialPad(
+      makeCall: makeCall,
+      keyPressed: keyPressed,
+      // Cupertino icons should be used here
+      dialButtonIcon: Icons.phone,
+      backspaceButtonIconColor: Colors.grey,
+      generator: PhoneKeypadGenerator(),
+      dialOutputTextColor: Colors.black87,
+      buttonTextColor: Colors.black54,
+      buttonColor: Colors.white,
+      buttonType: ButtonType.rectangle,
+      dialOutputTextSize: 75,
+      dialButtonColor: Colors.blue,
+    );
+  }
 
   @override
-  _DialPadState createState() => _DialPadState();
+  State<DialPad> createState() => _DialPadState();
 }
 
 class _DialPadState extends State<DialPad> {
-  MaskedTextController? textEditingController;
-  var _value = "";
-  var mainTitle = ["1", "2", "3", "4", "5", "6", "7", "8", "9", "*", "0", "＃"];
-  var subTitle = [
-    "",
-    "ABC",
-    "DEF",
-    "GHI",
-    "JKL",
-    "MNO",
-    "PQRS",
-    "TUV",
-    "WXYZ",
-    null,
-    "+",
-    null
-  ];
+  late final MaskedTextController _controller;
+  String _value = "";
 
   @override
   void initState() {
-    textEditingController = MaskedTextController(
-        mask: widget.outputMask != null ? widget.outputMask : '(000) 000-0000');
     super.initState();
+    _controller = MaskedTextController(mask: widget.outputMask);
   }
 
-  _setText(String? value) async {
-    if ((widget.enableDtmf == null || widget.enableDtmf!) && value != null)
-      Dtmf.playTone(digits: value.trim(), samplingRate: 8000, durationMs: 160);
+  /// Handles keypad button press, this includes numbers and [DialActionKey] except [DialActionKey.backspace]
+  void _onKeyPressed(String? value) {
+    if (value != null) {
+      setState(() {
+        _value += value;
+        _controller.updateText(_value);
+      });
+    }
+  }
 
-    if (widget.keyPressed != null) widget.keyPressed!(value!);
+  /// Handles backspace button press
+  void _onBackspacePressed() {
+    if (_value.isEmpty) {
+      return;
+    }
 
     setState(() {
-      _value += value!;
-      textEditingController!.text = _value;
+      _value = _value.substring(0, _value.length - 1);
+      _controller.text = _value;
     });
   }
 
-  List<Widget> _getDialerButtons() {
-    var rows = <Widget>[];
-    var items = <Widget>[];
-
-    for (var i = 0; i < mainTitle.length; i++) {
-      if (i % 3 == 0 && i > 0) {
-        rows.add(Row(
-            mainAxisAlignment: MainAxisAlignment.spaceEvenly, children: items));
-        rows.add(SizedBox(
-          height: 12,
-        ));
-        items = <Widget>[];
-      }
-
-      items.add(DialButton(
-        title: mainTitle[i],
-        subtitle: subTitle[i],
-        hideSubtitle: widget.hideSubtitle!,
-        color: widget.buttonColor,
-        textColor: widget.buttonTextColor,
-        onTap: _setText,
-      ));
+  /// Handles dial button press
+  void _onDialPressed() {
+    if (widget.makeCall != null && _value.isNotEmpty) {
+      widget.makeCall!(_value);
     }
-    //To Do: Fix this workaround for last row
-    rows.add(
-        Row(mainAxisAlignment: MainAxisAlignment.spaceEvenly, children: items));
-    rows.add(SizedBox(
-      height: 12,
-    ));
-
-    return rows;
   }
 
-  @override
-  Widget build(BuildContext context) {
-    var screenSize = MediaQuery.of(context).size;
-    var sizeFactor = screenSize.height * 0.09852217;
+  /// Handles keyboard button presses
+  void _onKeypadPressed(KeyValue key) {
+    if (key is ActionKey && key.action == DialActionKey.backspace) {
+      // handle backspace
+      _onBackspacePressed();
+    } else if (key is ActionKey && key.action == DialActionKey.enter) {
+      if (widget.callOnEnter) {
+        _onDialPressed();
+      }
+    } else {
+      // For numbers, and all actions except backspace
+      _onKeyPressed(key.value);
+    }
+  }
 
-    return Center(
-      child: Column(
-        children: <Widget>[
-          Padding(
-            padding: EdgeInsets.all(20),
-            child: TextFormField(
-              readOnly: true,
-              style: TextStyle(
-                  color: widget.dialOutputTextColor ?? Colors.black,
-                  fontSize: sizeFactor / 2),
-              textAlign: TextAlign.center,
-              decoration: InputDecoration(border: InputBorder.none),
-              controller: textEditingController,
-            ),
-          ),
-          ..._getDialerButtons(),
-          SizedBox(
-            height: 15,
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-            children: <Widget>[
-              Expanded(
-                child: Container(),
-              ),
-              Expanded(
-                child: widget.hideDialButton != null && widget.hideDialButton!
-                    ? Container()
-                    : Center(
-                        child: DialButton(
-                          icon: widget.dialButtonIcon != null ? widget.dialButtonIcon : Icons.phone,
-                          color: widget.dialButtonColor != null ? widget.dialButtonColor! : Colors.green,
-                          hideSubtitle: widget.hideSubtitle!,
-                          onTap: (value) {
-                            widget.makeCall!(_value);
-                          },
-                        ),
-                      ),
-              ),
-              Expanded(
-                child: Padding(
-                  padding:
-                      EdgeInsets.only(right: screenSize.height * 0.03685504),
-                  child: IconButton(
-                    icon: Icon(
-                      Icons.backspace,
-                      size: sizeFactor / 2,
-                      color: _value.length > 0
-                          ? (widget.backspaceButtonIconColor != null
-                              ? widget.backspaceButtonIconColor
-                              : Colors.white24)
-                          : Colors.white24,
-                    ),
-                    onPressed: _value.length == 0
-                        ? null
-                        : () {
-                            if (_value.length > 0) {
-                              setState(() {
-                                _value = _value.substring(0, _value.length - 1);
-                                textEditingController!.text = _value;
-                              });
-                            }
-                          },
-                  ),
-                ),
-              )
-            ],
-          )
-        ],
-      ),
+  Widget _defaultDialButtonBuilder(BuildContext context, int index, KeyValue key, KeyValue? altKey, String? hint) {
+    return DialButton(
+      title: key.value,
+      subtitle: altKey?.value ?? hint,
+      color: widget.buttonColor,
+      hideSubtitle: widget.hideSubtitle,
+      onTap: _onKeyPressed,
+      buttonType: widget.buttonType,
+      padding: widget.buttonPadding,
+      textColor: widget.buttonTextColor,
+      iconColor: widget.buttonTextColor,
+      subtitleIconColor: widget.buttonTextColor,
+      subtitleFontSize: widget.subtitleTextSize,
+      fontSize: widget.buttonTextSize,
     );
   }
-}
-
-class DialButton extends StatefulWidget {
-  final Key? key;
-  final String? title;
-  final String? subtitle;
-  final bool hideSubtitle;
-  final Color? color;
-  final Color? textColor;
-  final IconData? icon;
-  final Color? iconColor;
-  final ValueSetter<String?>? onTap;
-  final bool? shouldAnimate;
-  DialButton(
-      {this.key,
-      this.title,
-      this.subtitle,
-      this.hideSubtitle = false,
-      this.color,
-      this.textColor,
-      this.icon,
-      this.iconColor,
-      this.shouldAnimate,
-      this.onTap});
-
-  @override
-  _DialButtonState createState() => _DialButtonState();
-}
-
-class _DialButtonState extends State<DialButton>
-    with SingleTickerProviderStateMixin {
-  late AnimationController _animationController;
-  late Animation _colorTween;
-  Timer? _timer;
-
-  @override
-  void initState() {
-    _animationController =
-        AnimationController(vsync: this, duration: Duration(milliseconds: 300));
-    _colorTween = ColorTween(
-            begin: widget.color != null ? widget.color : Colors.white24,
-            end: Colors.white)
-        .animate(_animationController);
-
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    _animationController.dispose();
-    super.dispose();
-    if ((widget.shouldAnimate == null || widget.shouldAnimate!) &&
-        _timer != null) _timer!.cancel();
-  }
 
   @override
   Widget build(BuildContext context) {
-    var screenSize = MediaQuery.of(context).size;
-    var sizeFactor = screenSize.height * 0.09852217;
+    final screenSize = MediaQuery.of(context).size;
+    final sizeFactor = min(screenSize.height, screenSize.width) * 0.001;
 
-    return GestureDetector(
-      onTap: () {
-        if (this.widget.onTap != null) this.widget.onTap!(widget.title);
+    final _keypadButtonBuilder = /*widget.keypadButtonBuilder ??  */_defaultDialButtonBuilder;
+    final _generator = widget.generator ?? IosKeypadGenerator();
 
-        if (widget.shouldAnimate == null || widget.shouldAnimate!) {
-          if (_animationController.status == AnimationStatus.completed) {
-            _animationController.reverse();
-          } else {
-            _animationController.forward();
-            _timer = Timer(const Duration(milliseconds: 200), () {
-              setState(() {
-                _animationController.reverse();
-              });
-            });
-          }
-        }
-      },
-      child: ClipOval(
-          child: AnimatedBuilder(
-              animation: _colorTween,
-              builder: (context, child) => Container(
-                    color: _colorTween.value,
-                    height: sizeFactor,
-                    width: sizeFactor,
-                    child: Center(
-                        child: widget.icon == null
-                            ? widget.subtitle != null
-                                ? Column(
-                                    mainAxisSize: MainAxisSize.min,
-                                    children: <Widget>[
-                                      SizedBox(
-                                        height: 8,
-                                      ),
-                                      Text(
-                                        widget.title!,
-                                        style: TextStyle(
-                                            fontSize: sizeFactor / 2,
-                                            color: widget.textColor != null
-                                                ? widget.textColor
-                                                : Colors.black),
-                                      ),
-                                      if (!widget.hideSubtitle)
-                                      Text(widget.subtitle!,
-                                          style: TextStyle(
-                                              color: widget.textColor != null
-                                                  ? widget.textColor
-                                                  : Colors.black))
-                                    ],
-                                  )
-                                : Padding(
-                                    padding: EdgeInsets.only(
-                                        top: widget.title == "*" ? 10 : 0),
-                                    child: Text(
-                                      widget.title!,
-                                      style: TextStyle(
-                                          fontSize: widget.title == "*" &&
-                                                  widget.subtitle == null
-                                              ? screenSize.height * 0.0862069
-                                              : sizeFactor / 2,
-                                          color: widget.textColor != null
-                                              ? widget.textColor
-                                              : Colors.black),
-                                    ))
-                            : Icon(widget.icon,
-                                size: sizeFactor / 2,
-                                color: widget.iconColor != null
-                                    ? widget.iconColor
-                                    : Colors.white)),
-                  ))),
+    /// Dial button
+    final dialButton = widget.hideDialButton
+        ? null
+        : ActionButton(
+            padding: widget.buttonPadding,
+            buttonType: widget.buttonType,
+            icon: widget.dialButtonIcon,
+            iconColor: widget.dialButtonIconColor,
+            color: widget.dialButtonColor,
+            onTap: _onDialPressed,
+            // NOTE(cybex-dev) add as option in future
+            // disabled: _value.isEmpty || widget.makeCall == null,
+          );
+
+    /// Backspace button
+    final backspaceButton = widget.hideBackspaceButton
+        ? null
+        : ActionButton(
+            onTap: _onBackspacePressed,
+            disabled: _value.isEmpty,
+            buttonType: widget.buttonType,
+            iconSize: 75,
+            iconColor: widget.backspaceButtonIconColor,
+            padding: widget.buttonPadding,
+            icon: Icons.backspace,
+            color: Colors.transparent,
+          );
+
+    /// Footer contains the dial and backspace buttons
+    final footer = dialButton == null && backspaceButton == null
+        ? null
+        : Row(
+            children: [
+              Expanded(child: Container()),
+              Expanded(child: dialButton ?? Container()),
+              Expanded(child: backspaceButton ?? Container()),
+            ],
+          );
+
+    return KeypadFocusNode(
+      onKeypadPressed: _onKeypadPressed,
+      child: Center(
+        child: Column(
+          children: <Widget>[
+            Padding(
+              padding: widget.textFieldPadding,
+              child: PhoneTextField(
+                textStyle: TextStyle(
+                  color: widget.dialOutputTextColor,
+                  fontSize: widget.dialOutputTextSize * sizeFactor,
+                ),
+                decoration: InputDecoration(border: InputBorder.none, hintText: widget.hint),
+                controller: _controller,
+                copyToClipboard: widget.copyToClipboard,
+                readOnly: !widget.pasteFromClipboard,
+              ),
+            ),
+            Expanded(
+              child: KeypadGrid(
+                itemCount: 12,
+                itemBuilder: (context, index) {
+                  final key = _generator.get(index);
+                  final altKey = _generator.getAlt(index);
+                  final hint = _generator.hint(index);
+                  return _keypadButtonBuilder(context, index, key, altKey, hint);
+                },
+                footer: footer,
+              ),
+            )
+          ],
+        ),
+      ),
     );
   }
 }