+++ lib/src/flutter_dialpad.dart
@@ -1,0 +1,3 @@
+export 'generator/generator.dart';
+export 'models/models.dart';
+export 'widgets/widgets.dart';