+++ lib/src/widgets/buttons/button_type.dart
@@ -1,0 +1,4 @@
+enum ButtonType {
+  rectangle,
+  circle;
+}